# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Dare To Accept API",
    "version": "1.0.0",
    "description": "API documentation for Dare To Accept platform."
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/withdrawals": {
      "post": {
        "tags": [
          "Withdrawals"
        ],
        "summary": "Create a withdrawal request",
        "description": "Allows a user to request a withdrawal to their bank account.",
        "operationId": "createWithdrawal",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "amount": {
                  "type": "number",
                  "description": "Amount to withdraw"
                },
                "bankAccount": {
                  "type": "string",
                  "description": "Bank account details"
                }
              },
              "required": [
                "amount",
                "bankAccount"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Withdrawal request successful"
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Error processing withdrawal"
          }
        }
      }
    },
...
    "/challenges/{challengeId}/participants/{participantId}/out": {
      "post": {
        "tags": [
          "Challenges"
        ],
        "summary": "Admit defeat in a challenge",
        "description": "Allows a participant to admit defeat in a specific challenge.",
        "operationId": "admitDefeat",
        "parameters": [
          {
            "in": "path",
            "name": "challengeId",
            "required": true,
            "type": "string",
            "description": "The ID of the challenge."
          },
          {
            "in": "path",
            "name": "participantId",
            "required": true,
            "type": "string",
            "description": "The ID of the participant admitting defeat."
          }
        ],
        "responses": {
          "200": {
            "description": "Defeat admitted successfully",
            "schema": {
              "$ref": "#/definitions/Participant"
            }
          },
          "400": {
            "description": "Cannot admit defeat in current status"
          },
          "404": {
            "description": "Participant not found"
          }
        }
      }
    },
...
    "/api/withdrawals/pending": {
      "get": {
        "tags": [
          "Withdrawals"
        ],
        "summary": "Get all pending charity withdrawals",
        "description": "Returns a list of CharityPayouts with status 'pending'.",
        "operationId": "getPendingWithdrawals",
        "responses": {
          "200": {
            "description": "List of pending CharityPayouts",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CharityPayout"
              }
            }
          },
          "500": {
            "description": "Error fetching pending payouts"
          }
        }
      }
    },
    "/api/withdrawals/{id}/status": {
      "patch": {
        "tags": [
          "Withdrawals"
        ],
        "summary": "Update the status of a charity payout",
        "description": "Allows updating the status of a charity payout to 'sorted' or 'rejected'.",
        "operationId": "updatePayoutStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "The ID of the CharityPayout to update."
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "enum": [
                    "sorted",
                    "rejected"
                  ],
                  "description": "The new status of the payout."
                }
              },
              "required": [
                "status"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payout status updated successfully",
            "schema": {
              "$ref": "#/definitions/CharityPayout"
            }
          },
          "400": {
            "description": "Invalid status value"
          },
          "404": {
            "description": "Payout not found"
          },
          "500": {
            "description": "Error updating payout status"
          }
        }
      }
    }
  },
  "definitions": {
    "CharityPayout": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "Unique identifier for the CharityPayout."
        },
        "charity": {
          "type": "string",
          "description": "The ObjectId of the associated charity."
        },
        "charityName": {
          "type": "string",
          "description": "The name of the associated charity."
        },
        "amount": {
          "type": "number",
          "description": "The amount to be paid out."
        },
        "status": {
          "type": "string",
          "enum": [
            "pending",
            "sorted",
            "rejected"
          ],
          "description": "The current status of the payout."
        },
        "bankDetails": {
          "type": "object",
          "properties": {
            "bank_name": {
              "type": "string",
              "description": "The name of the bank."
            },
            "account_number": {
              "type": "string",
              "description": "The account number for the payout."
            },
            "routing_number": {
              "type": "string",
              "description": "The routing number for the payout."
            }
          }
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "description": "The timestamp when the payout was created."
        }
      }
    }
  },
...
}